refactor(api): split vault runtime modules
